+++ zcash_client_backend/src/data_api/wallet.rs
@@ -1431,7 +1431,7 @@ where
                 builder.add_orchard_output(
                     external_ovk.map(|k| k.into()),
                     to,
-                    payment_amount.into(),
+                    payment_amount,
                     memo.clone(),
                 )?;
                 orchard_output_meta.push((
@@ -1546,7 +1546,7 @@ where
                         ufvk.orchard()
                             .ok_or(Error::KeyNotAvailable(PoolType::ORCHARD))?
                             .address_at(0u32, orchard::keys::Scope::Internal),
-                        change_value.value().into(),
+                        change_value.value(),
                         memo.clone(),
                     )?;
                     orchard_output_meta.push((